parser: split up atoms -> variable/value
often used together, but two different animals nonetheless
