+++ src/parser/value.h
@@ -15,34 +15,37 @@
  * along with golog++.  If not, see <https://www.gnu.org/licenses/>.
 **************************************************************************/
 
-#ifndef GOLOGPP_PARSER_DOMAIN_H_
-#define GOLOGPP_PARSER_DOMAIN_H_
+#ifndef GOLOGPP_PARSER_VALUE_H_
+#define GOLOGPP_PARSER_VALUE_H_
 
 #include "utilities.h"
+#include "types.h"
 
-#include <model/domain.h>
 
 namespace gologpp {
 namespace parser {
 
-struct DomainExpressionParser : grammar<Domain(Scope &, Typename, bool)> {
-	DomainExpressionParser();
 
-	rule<Domain(Scope &, Typename, bool)> domain_expr;
-	rule <
-		Domain(Scope &, Typename, bool),
-		locals<shared_ptr<Domain>>
-	> unary_domain_expr;
-	rule<Domain(Scope &, Typename, bool)> binary_domain_expr;
-	rule<DomainOperator()> domain_operator;
-};
+rule<Value *()> &numeric_literal();
+rule<Value *()> &boolean_literal();
+rule<Value *()> &string_literal();
+rule<Value *()> &symbolic_literal();
+rule<Value *()> &symbolic_literal_def();
+rule<Value *()> &compound_literal();
+rule<Value *()> &list_literal();
+rule<Value *()> &undefined_literal();
 
-rule<Domain(Scope &, Typename)> &domain_expression();
-rule<void(Scope &), locals<string, Typename>> &domain_decl();
-rule<void(Scope &, bool), locals<Typename>> &domain_assignment();
+rule<Value *()> &any_literal();
+
+rule<Value *(Typename, bool)> &literal();
+
+
+void initialize_cyclic_literals();
 
 
 } // namespace parser
 } // namespace gologpp
 
-#endif //GOLOGPP_PARSER_DOMAIN_H_
+
+#endif // GOLOGPP_PARSER_VALUE_H_
+
